==> 1-data-collection-historical.ipynb <==
# # Configurar el ambiente de desarrollo
import pandas as pd
from urllib.parse import unquote, quote


# Extraer la data de la web
# Leer las páginas de Wikipedia que contienen los resultados de los juegos de La Liga en cada año

# Crear un diccionario con la temporada y la tabla de resultados
wiki = 'https://en.wikipedia.org/wiki/'
years = ['2017-18', '2018-19', '2019-20', '2020-21', '2021-22']
historical_data = {"season": [],
                   "data": []
                   }

for year in years:
    historical_data['season'].append(year)
    historical_data['data'].append(pd.read_html(f'{wiki}{year}_La_Liga')[5])

historical_data['data'][1]

# # Limpiar la data
# Agregar columna de temporada en historical_data y unificar el nombre de los equipos en los df
# Agregar una columna con la temporada en cada juego

for i in range(len(historical_data['data'])):
    season = historical_data['season'][i]
    data = historical_data['data'][i]
    data.insert(0, 'season', season)

# Cambiar los nombres de las columnas

for i in range(len(historical_data['data'])):
    short_names = historical_data['data'][i].columns[2:]
    long_names = historical_data['data'][i].iloc[:, 1]
    dict_team_names = dict(zip(short_names, long_names))

    # Renombrar las filas utilizando el diccionario generado long,shorts
    #historical_data['data'][i].replace(dict_team_names, inplace=True)
    # Renombrar las columnas utilizando el diccionario generado shorts,longs
    historical_data['data'][i].rename(columns=dict_team_names, inplace=True)

    # Establecer la primera fila como índice y cambiar el nombre de la columna "Unnamed: 0" por "Equipo"
    historical_data['data'][i] = historical_data['data'][i].set_index(
        historical_data['data'][i].columns[1]).rename_axis("Equipo")

historical_data['data'][1]

teams


historical_data['data'][4]

# Crear un nuevo dataframe que muestre los resultados en el formato deseado 'Local', 'Resultado', 'Visitante', 'Temporada
historical_results = pd.DataFrame(columns=['Local', 'Resultado', 'Visitante', 'Temporada', 'GolesLocal', 'GolesVisitante'])

for i in range(len(historical_data['data'])):
    historical_data_filtered = historical_data['data'][i].iloc[:, 0:]
    for index, row in historical_data_filtered.iterrows():
        for col_name, value in row.iloc[1:].items():
            # Obtener el resultado del partido en formato "goles_local - goles_visitante"
            score = value.strip()
            # Obtener el nombre del equipo visitante
            away_team = col_name
            # Obtener el nombre del equipo local
            local_team = index
            # Obtener la temporada
            season = historical_data['data'][i].iloc[0, 0]
            # Crear un nuevo dataframe con la fila actual
            new_row = pd.DataFrame({'Local': [local_team], 'Resultado': [score], 'Visitante': [away_team], 'Temporada': [season]})
            # Concatenar el nuevo dataframe con el dataframe de resultados existente
            historical_results = pd.concat([historical_results, new_row], ignore_index=True)

# Remover resultados entre los mismos equipos
historical_results = historical_results[historical_results['Local'] != historical_results['Visitante']]

# Separar la columna de resultado en goles local / goles visitante
historical_results[['GolesLocal', 'GolesVisitante']] = historical_results['Resultado'].str.split('–', expand=True)
historical_results.drop('Resultado', axis=1, inplace=True)

# Convertir las columnas de GolesLocal y GolesVisitante a tipo número
historical_results = historical_results.astype({'GolesLocal': int, 'GolesVisitante': int})

historical_results

from google.colab import drive
drive.mount('/content/drive')

pwd

cd ./drive/MyDrive/Colab Notebooks/laliga-nb

#Exportar a un archivo .csv
historical_results.to_csv('historical_results.csv',index=False)
