==> 2-data-collection-current-season.ipynb <==
## Configurar el ambiente de desarrollo
import pandas as pd
from urllib.parse import unquote, quote
from datetime import datetime


# Leer las tablas desde la URL
url = 'https://es.m.wikipedia.org/wiki/Primera_Divisi%C3%B3n_de_Espa%C3%B1a_2022-23'
all_tables = pd.read_html(url)

# Seleccionar solo las tablas que van del índice 14 al 52 menos la 19 que son las jornadas de la temporada 22-23
selected_tables = all_tables[14:53]
# Elimina la tabla de la posición 19
selected_tables.pop(19)

# Organizar la data
current_results = pd.DataFrame()
for table in selected_tables:
  # Eliminar el primer nivel del MultiIndex de las columnas
  table.columns = table.columns.droplevel(0)
  # Escoger solo las columnas que nos interesan
  table = table[['Local', 'Resultado', 'Visitante', 'Fecha']]
  current_results = pd.concat([current_results, table], ignore_index=True)

# Separar la columna de resultado en goles local / goles visitante
current_results[['GolesLocal', 'GolesVisitante']] = current_results['Resultado'].str.split('–', expand=True)
current_results.drop('Resultado', axis=1, inplace=True)

current_results

# Borrar las filas que tienen valores NaN
current_results = current_results.dropna()

# Convertir las columnas de GolesLocal y GolesVisitante a tipo número
current_results = current_results.astype({'GolesLocal': int, 'GolesVisitante': int})

current_results.tail

# Crear función para convertir la fecha
def date_conversion(date_str):
    month_dict = {
        'enero': 1,
        'febrero': 2,
        'marzo': 3,
        'abril': 4,
        'mayo': 5,
        'junio': 6,
        'julio': 7,
        'agosto': 8,
        'septiembre': 9,
        'octubre': 10,
        'noviembre': 11,
        'diciembre': 12
    }
    date_list = date_str.split()
    day = date_list[0]
    month_str = date_list[2]
    month = month_dict[month_str]
    if month <= 7:
        year = 2023
    else:
        year = 2022
    fecha = datetime(year, month, int(day))
    return fecha.strftime('%d/%m/%Y')

# Aplicar la función a la columna Fecha
current_results['Fecha'] = current_results['Fecha'].apply(date_conversion)

current_results['Fecha'] = pd.to_datetime(current_results['Fecha'], format='%d/%m/%Y')
current_results

current_results.dtypes


# Extraer el nombre limpio de los equipos que juegan la temporada 2022-23
wiki = 'https://en.wikipedia.org/wiki/'
years = '2022-23'
url = f'{wiki}{years}_La_Liga'

season_2223 = pd.read_html(url)[2]
teams_2223 = list(season_2223['Team'])

# Extraer el nombre a sustituir de los equipos del dataframe
teams_2223_to_change = current_results[['Local', 'Visitante']].values.flatten()
teams_2223_to_change = list(pd.unique(teams_2223_to_change))

# Arreglo con los nuevos nombres de los equipos
renaming_dict = {'C. A. Osasuna': 'Osasuna', 
                 'Sevilla F. C.': 'Sevilla',
                 'R. C. Celta de Vigo': 'Celta Vigo',
                 'R. C. D. Espanyol': 'Espanyol',
                 'Real Valladolid C. F.': 'Valladolid',
                 'Villarreal C. F.': 'Villarreal',
                 'F. C. Barcelona': 'Barcelona',
                 'Rayo Vallecano': 'Rayo Vallecano',
                 'Cádiz C. F.': 'Cádiz',
                 'Real Sociedad': 'Real Sociedad',
                 'Valencia C. F.': 'Valencia',
                 'Girona F. C.': 'Girona',
                 'U. D. Almería': 'Almería',
                 'Real Madrid C. F.': 'Real Madrid',
                 'Athletic Club': 'Athletic Bilbao',
                 'R. C. D. Mallorca': 'Mallorca',
                 'Getafe C. F.': 'Getafe',
                 'Atlético de Madrid': 'Atlético Madrid',
                 'Real Betis': 'Real Betis',
                 'Elche C. F.': 'Elche',
                 'Valencia F. C.': 'Valencia'}

# Sustituir los nombres en el dataframe
current_results.replace(renaming_dict, inplace=True)

# Arreglo con los nuevos nombres de los equipos
renaming_dict = {'C. A. Osasuna': 'Osasuna', 
                 'Sevilla F. C.': 'Sevilla',
                 'R. C. Celta de Vigo': 'Celta Vigo',
                 'R. C. D. Espanyol': 'Espanyol',
                 'Real Valladolid C. F.': 'Valladolid',
                 'Villarreal C. F.': 'Villarreal',
                 'F. C. Barcelona': 'Barcelona',
                 'Rayo Vallecano': 'Rayo Vallecano',
                 'Cádiz C. F.': 'Cádiz',
                 'Real Sociedad': 'Real Sociedad',
                 'Valencia C. F.': 'Valencia',
                 'Girona F. C.': 'Girona',
                 'U. D. Almería': 'Almería',
                 'Real Madrid C. F.': 'Real Madrid',
                 'Athletic Club': 'Athletic Bilbao',
                 'R. C. D. Mallorca': 'Mallorca',
                 'Getafe C. F.': 'Getafe',
                 'Atlético de Madrid': 'Atlético Madrid',
                 'Real Betis': 'Real Betis',
                 'Elche C. F.': 'Elche',
                 'Valencia F. C.': 'Valencia'}

current_results.replace(renaming_dict, inplace=True)


current_results

current_results = current_results.sort_values(by='Fecha', ascending=False)
current_results

from google.colab import drive
drive.mount('/content/drive')

pwd

cd ./drive/MyDrive/'Colab Notebooks'/laliga-nb

#Exportar a un archivo .csv
current_results.to_csv('current_results.csv',index=False)

# Crear una función para devolver los últimos N partidos de un equipo en particular

def last_matches_team(df, team, matches_num):
    # Filtrar los partidos del equipo especificado
    team_matches = df[(df['Local'] == team) | (df['Visitante'] == team)]
    
    # Ordenar los partidos por fecha descendente
    team_matches = team_matches.sort_values(by='Fecha', ascending=False)
    
    # Seleccionar los últimos matches_num partidos
    last_matches = team_matches.head(matches_num)
    
    # Devolver el DataFrame con los últimos partidos del equipo
    return last_matches


ultimos_partidos_barcelona =last_matches_team(current_results, 'Almería', 5)
print(ultimos_partidos_barcelona)
